feat(outlook-semantic-mcp): add MsGraphKqlSearchEmailsQuery with parallel KQL execution [UN-19787]
Create MsGraphKqlSearchEmailsQuery that accepts an array of raw KQL queries, executes them in parallel via Promise.all against the Graph /search/query endpoint, and deduplicates results by message ID. Move buildKqlQueryString, buildKqlForCondition, and the response Zod schemas into this new file; update GraphSearchEmailsQuery to import from there.
